+++ Backend/src/main/java/de/hsesslingen/scpprojekt/scp/Database/Controller/SportController.java
@@ -20,11 +20,11 @@ import java.util.Optional;
 /**
  * Rest Controller for Sport
  *
- * @author Tom Nguyen Dinh
+ * @author Tom Nguyen Dinh, Jason Patrick Duffy
  */
 @CrossOrigin(origins="http://localhost:3000", allowedHeaders = "*", allowCredentials = "true")
 @RestController
-@RequestMapping("/Sport")
+@RequestMapping("/sport")
 public class SportController {
 
     @Autowired
@@ -35,25 +35,25 @@ public class SportController {
      *
      * @param sport  sport data for the created one
      * @param request automatically filled by browser
-     * @return A 201 Code and the Challenge data if it worked 417 otherwise
+     * @return A 201 Code and the sport data if it worked 417 otherwise
      */
-    @Operation(summary = "Adds the new Challenge")
+    @Operation(summary = "Adds the new sport")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "201", description = "Sport successfully added",
                     content = {@Content(mediaType = "application/json",
                             schema = @Schema(implementation = Sport.class))}),
-            @ApiResponse(responseCode = "417", description = "Something went wrong creating the new sport", content = @Content),
+            @ApiResponse(responseCode = "500", description = "Something went wrong creating the new sport", content = @Content),
             @ApiResponse(responseCode = "403", description = "Not logged in", content = @Content)
     })
-    @PostMapping(path = "/addSport",produces = "application/json")
+    @PostMapping(path = "/add/",produces = "application/json")
     public ResponseEntity<Sport>addSport(@RequestBody Sport sport, HttpServletRequest request){
         if (SAML2Functions.isLoggedIn(request)){
             try {
                 Sport newsport = sportRepository.save(
-                        new Sport( sport.getName(),sport.getFactor()));
+                        new Sport(sport.getName(), sport.getFactor()));
                 return new ResponseEntity<>(newsport, HttpStatus.CREATED);
             }catch (Exception e){
-                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
             }
         } else {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
@@ -68,7 +68,7 @@ public class SportController {
      * @param request automatically filled by browser
      * @return Sport data corresponding to the given ID 404 otherwise
      */
-    @Operation(summary = "Get challenge by ID")
+    @Operation(summary = "Get sport by ID")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Sport found",
                     content = { @Content(mediaType = "application/json",
@@ -76,8 +76,8 @@ public class SportController {
             @ApiResponse(responseCode = "404", description = "Sport not found", content = @Content),
             @ApiResponse(responseCode = "403", description = "Not logged in", content = @Content)
     })
-    @GetMapping(path = "/{id}" , produces = "application/json")
-    public ResponseEntity<Sport> getChallengeById(@PathVariable("id") long id, HttpServletRequest request) {
+    @GetMapping(path = "/{id}/" , produces = "application/json")
+    public ResponseEntity<Sport> getSportById(@PathVariable("id") long id, HttpServletRequest request) {
         if (SAML2Functions.isLoggedIn(request)){
             Optional<Sport> sportData = sportRepository.findById(id);
             if (sportData.isPresent()) {
@@ -94,16 +94,16 @@ public class SportController {
      * Rest API for Returning all Sport
      *
      *  @param request automatically filled by browser
-     * @return Challenge data of all Challenges
+     * @return Sport data of all sports
      */
     @Operation(summary = "Get all sports")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "All sports that were found",
+            @ApiResponse(responseCode = "200", description = "Search successful",
                     content = {@Content(mediaType = "application/json",
                             array = @ArraySchema(schema = @Schema(implementation = Sport.class)))}),
             @ApiResponse(responseCode = "403", description = "Not logged in", content = @Content)
     })
-    @GetMapping(path = "/getAllSports", produces = "application/json")
+    @GetMapping(path = "/all/", produces = "application/json")
     public ResponseEntity<List<Sport>> getAllSports(HttpServletRequest request) {
         if (SAML2Functions.isLoggedIn(request)){
             List<Sport> sports = sportRepository.findAll();
@@ -118,28 +118,64 @@ public class SportController {
      *
      * @param id get SportID
      * @param request automatically filled by browser
-     * @return Signal for Deletion
+     * @return 204 for Deleted Sport if not 404
      */
     @Operation(summary = "Deleted the sport")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Deleted sport",
-                    content = {@Content(mediaType = "application/json",
-                            schema = @Schema(implementation = Challenge.class))}),
-            @ApiResponse(responseCode = "403", description = "Not logged in", content = @Content)
+            @ApiResponse(responseCode = "204", description = "Deleted sport"),
+            @ApiResponse(responseCode = "403", description = "Not logged in", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Sport not found", content = @Content)
     })
-    @DeleteMapping(path = "/{id}")
-    public ResponseEntity<HttpStatus> deleteSport(@PathVariable("id") long id, HttpServletRequest request) {
+    @DeleteMapping(path = "/")
+    public ResponseEntity<HttpStatus> deleteSport(@RequestParam("id") long id, HttpServletRequest request) {
         if (SAML2Functions.isLoggedIn(request)){
-            try {
+            Optional<Sport> sportData = sportRepository.findById(id);
+            if (sportData.isPresent()){
                 sportRepository.deleteById(id);
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-            } catch (Exception e) {
-                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            }else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
         } else {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             }
     }
+
+    /**
+     *Rest API for updating a sport
+     *
+     * @param ID of the sport which should be updated
+     * @param sport Sport data for Sport update
+     * @param request automatically filled by browser
+     * @return A 200 Code and the Sport data if it worked,  404 otherwise
+     */
+    @Operation(summary = "Updates a sport")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "sport successfully updated",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Sport.class))}),
+            @ApiResponse(responseCode = "404", description = "Sport not found", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Not logged in", content = @Content)
+    })
+    @PutMapping(path = "/", produces = "application/json")
+    public ResponseEntity<Sport> updateChallenge(@RequestParam("id") Long ID, @RequestBody Sport sport, HttpServletRequest request) {
+        if (SAML2Functions.isLoggedIn(request)){
+            Optional<Sport> sportData = sportRepository.findById(ID);
+
+            if (sportData.isPresent()) {
+                Sport newSport = sportData.get();
+                newSport.setName(sport.getName());
+                newSport.setFactor(sport.getFactor());
+
+                return new ResponseEntity<>(sportRepository.save(newSport), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+    }
+
 
 
 }
